backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.6.0
motor==3.3.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0